moved, renamed LZ4Utils
